backend/pyproject.toml
[project]
name = "perpetua-flow-backend"
version = "0.1.0"
description = "Backend API for Perpetua Flow task management application"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Perpetua Flow Team"}
]

dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",

    # Database
    "sqlmodel>=0.0.14",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "alembic>=1.13.1",

    # Authentication
    "pyjwt[crypto]>=2.8.0",
    "cryptography>=42.0.0",
    "cachetools>=5.3.0",

    # HTTP Client
    "httpx>=0.26.0",

    # Date/Time
    "python-dateutil>=2.8.2",

    # AI Integration
    "openai-agents>=0.0.7",
    "deepgram-sdk>=3.1.0",

    # Rate Limiting
    "slowapi>=0.1.9",

    # Configuration
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",

    # Metrics & Observability
    "prometheus-client>=0.19.0",
    "structlog>=24.1.0",

    # Validation
    "email-validator>=2.1.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "factory-boy>=3.3.0",
    "respx>=0.20.2",
    "schemathesis>=3.25.0",

    # Code Quality
    "ruff>=0.1.14",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",

    # Database Testing
    "pytest-postgresql>=6.0.0",
    "aiosqlite>=0.19.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "contract: Contract tests",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
